Moved sources and preferences to main directory.
