+++ packaging/configurer/configurer_yum.go
@@ -4,18 +4,22 @@
 
 package configurer
 
+import (
+	"github.com/juju/utils/packaging"
+)
+
 // yumConfigurer is the PackagingConfigurer implementation for apt-based systems.
 type yumConfigurer struct {
 	*baseConfigurer
 }
 
 // RenderSource implements PackagingConfigurer.
-func (c *yumConfigurer) RenderSource(src PackageSource) string {
-	return src.renderSourceFile(YumSourceTemplate[1:])
+func (c *yumConfigurer) RenderSource(src packaging.PackageSource) string {
+	return src.RenderSourceFile(YumSourceTemplate[1:])
 }
 
 // RenderPreferences implements PackagingConfigurer.
-func (c *yumConfigurer) RenderPreferences(src PackagePreferences) string {
+func (c *yumConfigurer) RenderPreferences(src packaging.PackagePreferences) string {
 	// TODO (aznashwan): research a way of using yum-priorities in the context
 	// of single/multiple package pinning and implement it.
 	return ""
